requirements.txt
utils-nuuuwan
gig-nuuuwan
camelot-py[cv]
matplotlib
